Update README.md
Thought it will be good idea to include `JsonLowerCaseNamingPolicy` in example, started to look for it in google and then realized it's a custom policy. Maybe also good to include this naming policy in source code so folks don't need to create it? Let me know, happy to do it.